backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
mysqlclient==2.2.0
Pillow==10.1.0
django-filter==23.3
python-decouple==3.8
celery==5.3.4
redis==5.0.1
reportlab==4.0.7
openpyxl==3.1.2
django-extensions==3.2.3